Add files via upload
//the task is following:
/*Write a program to read an arbitrary-length list of scores between 1 and 99 from a text file named
scores.txt. After reading the scores, your program must analyze all of them, determine the average score
and print statistics about the number of scores less than or equal to that average and the number of scores
above it. The following is a sample output of the program:
*/
//main working principle:
/*I am going to calculate sum at first and then calculate the average.
I dont know how to create an int array that will be infinite, therefore, after calculating the mean,
I am going to re-run the loop and compare each score to the mean.

I made this code so that in the file, it should only be the numbers in the specific order. 
one line should contain only one score and nothing else (except'\n') 
I made this on purpose to be neat*/